FEAT: 캐릭터 이동 시 좌표 double로 변경

+++ backend/arc205/src/main/java/com/project/arc205/gamecharacter/dto/Location.java
@@ -12,5 +12,5 @@ import java.io.Serializable;
 @AllArgsConstructor
 @ToString
 public class Location implements Serializable {
-    private int y, x;
+    private Double y, x;
 }
